refactor(vscode): replace View Diff with Open Worktree in New Window
The previous View Diff command tried to render a multi-file diff in
the current VSCode window via `vscode.changes` + `git:` URIs. It
silently failed — title showed "(2 files)" but the body said "No
Changed Files" — because VSCode's Git extension doesn't auto-discover
worktrees that live inside the host workspace's gitignore. Tried
fixing URI shape (absolute path in query, ref-based instead of
empty-ref); didn't help because the Git extension wasn't looking at
the worktree's git database in the first place.

The right shape is a custom TextDocumentContentProvider that resolves
content via `git -C <worktree> show <ref>:<path>` ourselves —
bypassing the Git extension entirely. That's a fair chunk of code and
a separate concern; meanwhile reviewers need something that works.

Replace View Diff with **Open Worktree in New Window**:
`vscode.commands.executeCommand('vscode.openFolder', uri, true)` opens
a new editor window rooted at `.builders/<id>/`. In that window the
Git extension treats the worktree as a primary checkout — SCM panel,
working-tree changes, inline diffs against the branch base all work
natively.

`vscode.openFolder` is editor-agnostic — VSCode, Cursor, Windsurf,
and other VSCode-family editors all expose it identically. No CLI
detection needed.

Trade-off: review happens in a separate window instead of inline.
Acceptable for now given that (a) the inline path didn't work, and
(b) the new-window UX is closer to how reviewers actually work
(focused on the worktree, full SCM panel available).

Removed:
- packages/vscode/src/commands/review-diff.ts (~180 LOC)
- `codev.reviewDiff` command declaration and menu entry

Added:
- packages/vscode/src/commands/open-worktree-window.ts (~75 LOC)
- `codev.openWorktreeWindow` declaration + menu entry in slot
  3_review@1 (replaces View Diff's slot)
